+++ packages/datepicker-suggest-core/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @datepicker-suggest/core
 
+## 0.0.5
+
+### Patch Changes
+
+- fixes for first integration
+
 ## 0.0.4
 
 ### Patch Changes